chore: additional "add" helpers (#4840)
* more add helpers

* add duplicate
