+++ lib/store/helpers/add/index.ts
@@ -1,3 +1,7 @@
 import { add } from "./add";
 import { addSubItem } from "./addSubItem";
-export { add, addSubItem };
+import { addChoice } from "./addChoice";
+import { addLabeledChoice } from "./addLabeledChoice";
+import { addSubChoice } from "./addSubChoice";
+import { duplicateElement } from "./duplicateElement";
+export { add, addSubItem, addChoice, addLabeledChoice, addSubChoice, duplicateElement };
