+++ lib/store/useTemplateStore.tsx
@@ -37,7 +37,14 @@ import { clearTemplateStorage } from "./utils";
 import { orderGroups } from "@lib/utils/form-builder/orderUsingGroupsLayout";
 import { initStore } from "./initStore";
 
-import { add, addSubItem } from "./helpers/add";
+import {
+  add,
+  addSubItem,
+  addChoice,
+  addLabeledChoice,
+  addSubChoice,
+  duplicateElement,
+} from "./helpers/add";
 import { moveUp, moveDown, subMoveUp, subMoveDown } from "./helpers/move";
 
 const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => {
@@ -114,11 +121,15 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
               set((state) => {
                 unset(state, path);
               }),
-            moveUp: moveUp(set, get),
-            moveDown: moveDown(set, get),
-            subMoveUp: subMoveUp(set, get),
-            subMoveDown: subMoveDown(set, get),
+            moveUp: moveUp(set),
+            moveDown: moveDown(set),
+            subMoveUp: subMoveUp(set),
+            subMoveDown: subMoveDown(set),
             add: add(set, get),
+            addSubItem: addSubItem(set, get),
+            addChoice: addChoice(set),
+            addLabeledChoice: addLabeledChoice(set),
+            addSubChoice: addSubChoice(set),
             removeChoiceFromRules: (elId: string, choiceIndex: number) => {
               set((state) => {
                 const choiceId = `${elId}.${choiceIndex}`;
@@ -138,7 +149,6 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
                 state.form.groups = groups;
               });
             },
-            addSubItem: addSubItem(set, get),
             remove: (elementId, groupId = "") => {
               set((state) => {
                 const allowGroups = state.allowGroupsFlag;
@@ -169,32 +179,6 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
                   );
                 }
               }),
-            addChoice: (elIndex) =>
-              set((state) => {
-                state.form.elements[elIndex].properties.choices?.push({ en: "", fr: "" });
-              }),
-            addLabeledChoice: async (elIndex, label) => {
-              return new Promise((resolve) => {
-                set((state) => {
-                  state.form.elements[elIndex].properties.choices?.push({
-                    en: label.en,
-                    fr: label.fr,
-                  });
-
-                  const lastChoice = state.form.elements[elIndex].properties.choices?.length ?? 0;
-                  resolve(lastChoice);
-                });
-              });
-            },
-            addSubChoice: (elId, subIndex) => {
-              set((state) => {
-                const parentIndex = getParentIndex(elId, state.form.elements);
-                if (parentIndex === undefined) return;
-                state.form.elements[parentIndex].properties.subElements?.[
-                  subIndex
-                ].properties.choices?.push({ en: "", fr: "" });
-              });
-            },
             removeChoice: (elIndex, choiceIndex) =>
               set((state) => {
                 state.form.elements[elIndex].properties.choices?.splice(choiceIndex, 1);
@@ -211,40 +195,7 @@ const createTemplateStore = (initProps?: Partial<InitialTemplateStoreProps>) => 
               const elIndex = get().form.elements.findIndex((el) => el.id === elId);
               return get().form.elements[elIndex]?.properties.choices?.[choiceIndex];
             },
-            duplicateElement: (itemId, groupId = "", copyEn = "", copyFr = "") => {
-              const elIndex = get().form.elements.findIndex((el) => el.id === itemId);
-              const id = get().generateElementId();
-              set((state) => {
-                // deep copy the element
-                const element = JSON.parse(JSON.stringify(state.form.elements[elIndex]));
-                element.id = id;
-                if (element.type !== "richText" && element.properties["titleEn"]) {
-                  element.properties["titleEn"] = `${element.properties["titleEn"]} ${copyEn}`;
-                }
-                if (element.type !== "richText" && element.properties["titleFr"]) {
-                  element.properties["titleFr"] = `${element.properties["titleFr"]} ${copyFr}`;
-                }
-                state.form.elements.splice(elIndex + 1, 0, element);
-                state.form.layout.splice(elIndex + 1, 0, id);
-
-                // Handle groups
-                const allowGroups = state.allowGroupsFlag;
-                groupId = allowGroups && groupId ? groupId : "";
-                if (allowGroups && groupId) {
-                  if (!state.form.groups) state.form.groups = {};
-                  if (!state.form.groups[groupId])
-                    state.form.groups[groupId] = {
-                      name: "",
-                      titleEn: "",
-                      titleFr: "",
-                      elements: [],
-                    };
-                  state.form.groups &&
-                    state.form.groups[groupId].elements.splice(elIndex + 1, 0, String(id));
-                }
-                // end groups
-              });
-            },
+            duplicateElement: duplicateElement(set, get),
             getSchema: () => {
               return JSON.stringify(getSchemaFromState(get(), get().allowGroupsFlag), null, 2);
             },
